move non-operations from intranetSchema to private
