+++ codegen.json
@@ -14,13 +14,14 @@
       "plugins": ["typescript", "typescript-operations", "add"]
     },
     "packages/server/graphql/private/resolverTypes.ts": {
-      "schema": "packages/server/graphql/intranetSchema/intranetSchema.graphql",
+      "schema": "packages/server/graphql/private/schema.graphql",
       "plugins": ["typescript", "typescript-resolvers", "add"],
       "config": {
         "contextType": "../../graphql#InternalContext",
         "mappers": {
           "PingableServices": "./types/PingableServices#PingableServicesSource",
-          "LoginsPayload": "./types/LoginsPayload#LoginsPayloadSource"
+          "LoginsPayload": "./types/LoginsPayload#LoginsPayloadSource",
+          "SignupsPayload": "./types/SignupsPayload#SignupsPayloadSource"
         }
       }
     }
