+++ packages/server/graphql/executeGraphQL.ts
@@ -11,7 +11,7 @@ import PROD from '../PROD'
 import CompiledQueryCache from './CompiledQueryCache'
 import getDataLoader from './getDataLoader'
 import getRateLimiter from './getRateLimiter'
-import privateSchema from './intranetSchema/intranetSchema'
+import privateSchema from './private/rootSchema'
 import publicSchema from './rootSchema'
 
 export interface GQLRequest {
