+++ packages/server/graphql/private/mutations/updateEmail.ts
@@ -1,7 +1,7 @@
-import getRethink from '../../../../database/rethinkDriver'
-import {getUserByEmail} from '../../../../postgres/queries/getUsersByEmails'
-import updateUser from '../../../../postgres/queries/updateUser'
-import {requireSU} from '../../../../utils/authorization'
+import getRethink from '../../../database/rethinkDriver'
+import {getUserByEmail} from '../../../postgres/queries/getUsersByEmails'
+import updateUser from '../../../postgres/queries/updateUser'
+import {requireSU} from '../../../utils/authorization'
 import {MutationResolvers} from '../resolverTypes'
 
 const updateEmail: MutationResolvers['updateEmail'] = async (
