+++ packages/server/graphql/private/queries/logins.ts
@@ -1,6 +1,5 @@
 import {QueryResolvers} from '../resolverTypes'
 
-
 const logins: QueryResolvers['logins'] = async (_source, args) => args
 
 export default logins
