+++ packages/server/graphql/private/queries/ping.ts
