+++ packages/server/graphql/private/typeDefs/ping.graphql
