+++ packages/server/graphql/private/types/LoginsPayload.ts
@@ -1,17 +1,19 @@
-import authCount from "../../intranetSchema/queries/helpers/authCount"
-import authCountByDomain from "../../intranetSchema/queries/helpers/authCountByDomain"
-import { LoginsPayloadResolvers } from '../resolverTypes'
+import authCount from '../../intranetSchema/queries/helpers/authCount'
+import authCountByDomain from '../../intranetSchema/queries/helpers/authCountByDomain'
+import {LoginsPayloadResolvers} from '../resolverTypes'
 
-export type LoginsPayloadSource = any
+export interface LoginsPayloadSource {
+  after?: Date
+  isActive: boolean
+}
 
 const LoginsPayload: LoginsPayloadResolvers = {
-    total: async ({after, isActive}) => {
-      return authCount(after, isActive, 'lastSeenAt')
-    },
-
-    byDomain: async ({after, isActive}) => {
-      return authCountByDomain(after, isActive, 'lastSeenAt')
-    }
+  total: async ({after, isActive}) => {
+    return authCount(after, isActive, 'lastSeenAt')
+  },
+  byDomain: async ({after, isActive}) => {
+    return authCountByDomain(after, isActive, 'lastSeenAt')
+  }
 }
 
 export default LoginsPayload
