+++ packages/server/graphql/private/types/SignupsPayload.ts
@@ -1,17 +1,20 @@
-import authCount from "../../intranetSchema/queries/helpers/authCount"
-import authCountByDomain from "../../intranetSchema/queries/helpers/authCountByDomain"
-import { SignupsPayloadResolvers } from '../resolverTypes'
+import authCount from '../../intranetSchema/queries/helpers/authCount'
+import authCountByDomain from '../../intranetSchema/queries/helpers/authCountByDomain'
+import {SignupsPayloadResolvers} from '../resolverTypes'
 
-export type SignupsPayloadSource = any
+export type SignupsPayloadSource = {
+  after: Date | null
+  isActive: boolean
+}
 
 const SignupsPayload: SignupsPayloadResolvers = {
-    total: async ({after, isActive}) => {
-      return authCount(after, isActive, 'createdAt')
-    },
+  total: async ({after, isActive}) => {
+    return authCount(after, isActive, 'createdAt')
+  },
 
-    byDomain: async ({after, isActive}) => {
-      return authCountByDomain(after, isActive, 'createdAt')
-    }
+  byDomain: async ({after, isActive}) => {
+    return authCountByDomain(after, isActive, 'createdAt')
+  }
 }
 
 export default SignupsPayload
