+++ packages/server/tsconfig.eslint.json
@@ -8,7 +8,7 @@
     "postgres/migrations/postDeploy/*",
     "postgres/types/*",
     "graphql/queries",
-    "graphql/intranetSchema/sdl/**/*.ts",
+    "graphql/private/**/*.ts",
     "integrations",
     "email/**/*",
     "segmentFns/*",
