+++ packages/server/utils/updateGQLSchema.ts
@@ -2,8 +2,8 @@ import fs from 'fs'
 import {printSchema} from 'graphql'
 import path from 'path'
 import {promisify} from 'util'
-import privateSchema from '../graphql/intranetSchema/intranetSchema'
-import schema from '../graphql/rootSchema'
+import privateSchema from '../graphql/private/rootSchema'
+import publicSchema from '../graphql/rootSchema'
 
 declare const __PROJECT_ROOT__: string
 
@@ -11,10 +11,10 @@ const write = promisify(fs.writeFile)
 const schemaPath = path.join(__PROJECT_ROOT__, 'schema.graphql')
 const privateSchemaPath = path.join(
   __PROJECT_ROOT__,
-  'packages/server/graphql/intranetSchema/intranetSchema.graphql'
+  'packages/server/graphql/private/schema.graphql'
 )
 const updateGQLSchema = async () => {
-  const nextSchema = printSchema(schema)
+  const nextSchema = printSchema(publicSchema)
   const nextPrivateSchema = printSchema(privateSchema)
   await Promise.all([write(schemaPath, nextSchema), write(privateSchemaPath, nextPrivateSchema)])
 }
