+++ pm2.dev.config.js
@@ -6,7 +6,7 @@ module.exports = {
       instances: 1,
       autorestart: true,
       watch: ['packages/server', 'packages/gql-executor'],
-      ignore_watch: ['**/__tests__', '**/intranetSchema.graphql'],
+      ignore_watch: ['**/__tests__', '**/rootSchema.graphql'],
       max_memory_restart: '3000M',
       env_production: {
         NODE_ENV: 'development'
